feat: Catch auth session between requests
The auth module is now cached and reused for folloup registry requests.

Previously each request to registry initialized a new session which lead
to reset the access token and need of establish secret session all over
again. This is now fixed and leads to huge performance improvement when
making multiple queries to a registry.

JIRA: ISV-4429

Signed-off-by: Ales Raszka <araszka@redhat.com>